% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_analytic.r
\name{channel_analytic}
\alias{channel_analytic}
\title{channel_analytic}
\usage{
channel_analytic(channel_obj, use_channel_dates = TRUE, start_date = NULL,
  end_date = NULL, Ntop = 11, Nmin = 25, naming = "",
  only_original_tweet = FALSE, lowercase = TRUE, stopword = "ITA",
  account_tw = "", corpus_hashtag = TRUE, graph_analisys = FALSE,
  corpus_analisys = FALSE)
}
\arguments{
\item{channel_obj}{data.frame Dataset of tweets}

\item{use_channel_dates}{Logical Use temporal indication of channel}

\item{start_date}{Character Date of analisys starting.}

\item{end_date}{Character Date of  analisys ending.}

\item{Ntop}{Integer indicate the maximum number for top statistics}

\item{Nmin}{Integer indicate the minimal data numerosity}

\item{naming}{Character Indicate which naming framework is adopted.}

\item{only_original_tweet}{Logical Taking into account only original. Default all tweets are considered.}

\item{lowercase}{logical Consider  all text as lower case. Default is TRUE.}

\item{stopword}{Character stopword set to be use to calculate word frequency matrix. Default italian stopwords of R tm package.}

\item{account_tw}{CharacterUser account if naming parameter is an "account_statistics".}

\item{corpus_hashtag}{logical Corpus analisys not considering the hashtag.}

\item{graph_analisys}{Igraph object Graph analisys done. default is FALSE.}

\item{corpus_analisys}{qDap object Corpora analisys done. Default is FALSE.}
}
\value{
Return a R list object  for channel analisys:

channel_stat list: Channel summaries of following parameters.

daily_stat data.frame : Daily channel statistics featured by

table_message data.frame :Frequency data  of messages.

table_hash data.frame : Frequency data  hashtags.

table_mentions data.frame : Frequency data of mentions.

table_links data.frame : Frequency data.frame of urls.

table_authors data.frame : Frequency data of authors.

table_authors_retweeter data.frame : Frequency data.frame of authors that are active retweeter.

retweeted_authors data.frame : Frequency of retweeted authors.

favorite_authors data.frame : Frequency data of favorited authors.

favorite_message data.frame : Rank of Favorite messages.

rank_message_retweeted data.frame : Frequency of retweeted message.

top_message data.frame : TopN messages in channel.

top_authors data.frame : TopN authors in channel.

top_hash data.frame : TopN hashtag.

top_mentions data.frame : TopN user mentioned.

top_links data.frame : TopN links.

topfull_authors_retweeter data.frame : TopN author that have retweeted.

topfull_message_retweeted data.frame : TopN message that have retweeted-

graph_retweet igraph object: Simplified Retweet graph

graph_mentions igraph object: Simplified Mention graph object as igraph R object

channel_data data.frame: Channel_data used for analisys with other ancillary variables.

channel_corpus wfm object:  Corpus of messages without mentions and links and optionally without hashtag.

word_freq_matr qdap wfm object : Word frequency matrix.

account_stats data.frame : Statistic account's activity by date.
}
\description{
Extract many informative stats and object  from a set of tweet messages parsed as channel
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{channel,stats}

