% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/channel_annotation_names.r
\name{channel_annotation_names}
\alias{channel_annotation_names}
\title{channel_annotation_names}
\usage{
channel_annotation_names(channel = "allertameteoLIG", target = "hashtag",
  repo = "https://raw.github.com/alfcrisci/twchannels_notations/master/")
}
\arguments{
\item{channel}{character  Name of channel in repository}

\item{target}{character  Target of annotation ( user,hashtag, geonames or other)}

\item{repo}{character  Address of repository where annotation key file are hosted.}
}
\value{
x    data.frame   A data.frame with key and category
}
\description{
Retrieve the names of categorical annotations from a github public repository related to annotation regarding the channel.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotations}
\keyword{names,}

