% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/channel_outputs.r
\name{channel_outputs}
\alias{channel_outputs}
\title{channel_outputs}
\usage{
channel_outputs(stat_obj, param = "channel_stat", suffix_file = "LIG",
  na_string = "", filecsv = FALSE, html = TRUE, excel = TRUE,
  corpus_hashtag = FALSE)
}
\arguments{
\item{stat_obj}{list  output obtained by channel_analytic}

\item{param}{character Name of statistics or object of channel_analytic object.}

\item{suffix_file}{character Prefix for file names outputs}

\item{na_string}{character String value for missing data.}

\item{filecsv}{logical export csv files.}

\item{html}{logical export html sjPlot files.}

\item{excel}{logical export excel files.}

\item{corpus_hashtag}{logical Corpus not taking into account hashtag.}
}
\value{
Object extracted from the list of output given by \code{channel_analytic} object.
}
\description{
Export result from statistical of \code{channel_analytic} function outputs.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotation}

