% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_unicode.r
\name{clean_unicode}
\alias{clean_unicode}
\title{clean_unicode}
\usage{
clean_unicode(variable)
}
\arguments{
\item{variable}{character  char vector}
}
\value{
Return a text vector cleaned
}
\description{
Wrapper function to provide elimination of unicode characters in text array
}
\author{
NYU Center for Data Science Pablo Barberá \email{pablo.barbera@nyu.edu}
}
\keyword{unicode}

