% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_stat_note_obj.r
\name{create_state_obj}
\alias{create_state_obj}
\title{create_state_obj}
\usage{
create_state_obj(x, notes, save = TRUE, fileName = "")
}
\arguments{
\item{x}{data.frame  Terms or user date list object from \code{channel_analytics}}

\item{qdap_hash}{Annotation file}
}
\value{
list Return stat object
}
\description{
Create a stat note object from time list of hashtag
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{match}

