% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_xts_from_df.r
\name{create_xts_from_df}
\alias{create_xts_from_df}
\title{create_xts_from_df}
\usage{
create_xts_from_df(x, ntimevar)
}
\arguments{
\item{x}{data.frame  Data.frame with time var}

\item{ntimevar}{numeric  Index of var where time is indexed}
}
\value{
A xts time series object.
}
\description{
Wrapper function to provide xts object creation from a data.frame object
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{xts,timeseries}

