% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_channel_stat.r
\name{daily_channel_stat}
\alias{daily_channel_stat}
\title{daily_channel_stat}
\usage{
daily_channel_stat(x, stream = "")
}
\arguments{
\item{x}{Data.frame Dataset of tweets}

\item{stream}{Character Type of stream.}
}
\value{
daily_channel_final Data.frame of daily stats performed by \code{fastChannelstat}
}
\description{
Function to compute daily summary stats of a twitter channel.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{channel,}
\keyword{channel.}
\keyword{stats,}
\keyword{twitter}
