% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_channel.r
\name{download_channel}
\alias{download_channel}
\title{download_channel}
\usage{
download_channel(channel, outfile, start_date, end_date, user, pass,
  format = "csv")
}
\arguments{
\item{channel}{Character  Name of channel in web format ( without blank space)-}

\item{outfile}{Character  Name of file to be saved}

\item{start_date}{Character Date of end of channel (format YYYY-MM-DD).}

\item{end_date}{Character  Date of end of channel (format YYYY-MM-DD).}

\item{user}{Character  User for disit platform. Required.}

\item{pass}{Character  Password for disit platform. Required.}

\item{format}{Character  File format for downloaded file. Default is csv.}
}
\value{
File of channel data.
}
\description{
To be used to download channel data from http://disit.org/tv/ Twitter Vigilance of University of Florence.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{DISIT,twitter}
\keyword{vigilance}

