% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_channel.r
\name{export_data_channel}
\alias{export_data_channel}
\title{export_data_channel}
\usage{
export_data_channel(x, name, native = F, graph = T, corpus = F,
  excel = TRUE, csv = FALSE, filterdegree = 2, saveres = F,
  export_graph = F, format_graph = "graphml")
}
\arguments{
\item{x}{Data.frame Channel anlitics.}

\item{native}{logical If data are data native where retweet are not considered.}

\item{graph}{logical It works on graph data.}

\item{corpus}{logical It exports word frequency matrix.}

\item{excel}{logical It exports excel.}

\item{csv}{logical It exports csv.}

\item{filterdegree}{numeric Filter level of graph.}

\item{export_graph}{logical It exports graph.}

\item{format_graph}{character Format of graph exported.Default is graphml.}

\item{anme}{character Label of stream.}

\item{saaveres}{logical It exports results in list.}
}
\value{
List and exports file
}
\description{
Function to export informative stats  from twitter analytic channel.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{,}
\keyword{channel}
\keyword{stats}

