% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hashtag.r
\name{extract_hashtag}
\alias{extract_hashtag}
\title{extract_hashtag}
\usage{
extract_hashtag(x)
}
\arguments{
\item{x}{Character  Messages of channel's tweets.}
}
\value{
Return the list of mentioned users
}
\description{
Extract hashtag mentioned in a message of tweet.
}
\references{
qdap R packages
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mentions}
