% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastChannelstat.r
\name{fastChannelstat}
\alias{fastChannelstat}
\title{fastChannelstat}
\usage{
fastChannelstat(x, check_duplicates = FALSE, stream = "")
}
\arguments{
\item{x}{Data.frame Channel time streams with tweets}

\item{check_duplicates}{Logical Check if exist duplicated tweets.}

\item{stream}{Character Kind of channel tweets collection. Default is empty.}
}
\value{
RTW_TW Numeric Number of full tweets ( tweets and retweets).

TW Numeric Number of original tweets.

RTW Numeric Number of original retweets.

ratioRTW_TW Numeric Ratio full ( tweets and retweets) vs original tweets.

ini_date_full  Date channel initial.

end_date_full Date channel final.

activity_days Numeric day of activity channel.

activity_days_native  Numeric day of activity channel.

period_extent Numeric Days covering channle period.

relative_activity Numeric Full activity ratio.

RTW_TW_daily  Numeric Full number of tweets.

TW_daily  Numeric Native number of tweets.

retweetCount  Numeric  Sum of retweets.

favoriteCount  Numeric Sum of favorites.

U_full_users  Numeric Number of full unique users.

N_full_hashtag  Numeric  Number of hashtag used.

U_full_hashtag  Numeric  Number of unique hashtag used.

N_full_mentions Numeric Number of mentions.

U_full_mentions  Numeric Number of unique mentions.

N_full_links  Numeric  Number of links.

U_full_links  Numeric  Number of unique links.

mostRT_msg Character Most retweeted message  ( tweets or retweets).

U_native_users  Numeric  Number of native unique users.

N_native_hashtag  Numeric  Number of hashtag used.

U_native_hashtag  Numeric  Number of unique hashtag used.

N_native_mentions  Numeric Number of mentions.

U_native_mentions  Numeric Number of unique mentions.

N_native_links  Numeric Number of links.

U_native_links  Numeric Number of native unique users.

mostRT_msg_native Character Most retweeted message native.

M_ch_counts_native Numeric Mean Number of original tweets favorite.

N_replies Numeric Number of original tweets favorite.

M_ch_counts_full Numeric mean

most_favorited_messages Character Most favorited message.

most_mentioned Character Most mentioned user.

most_retweeted Character Most mentioned user.

most_favorited Character

N_favor_full Numeric Number of original tweets favorite.

Nfavor_native Numeric Number of original tweets favorite.

N_geo Numeric Number of Geolocated tweets.
}
\description{
Function working on channel extracting many informative stats  from twitter timelines  parsed as channel.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{channel,daily}
\keyword{stats}
