% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorpus.r
\name{getCorpus}
\alias{getCorpus}
\title{getCorpus}
\usage{
getCorpus(textVector, hashtag = FALSE, elim_words)
}
\arguments{
\item{textVector}{Character  Message of tweet}

\item{hashtag}{logical remove hashtag from corpus}

\item{elim_words}{character the words to be removed from corpus}
}
\value{
Return   a tm corpus of message purged by mentions and links.
}
\description{
Creating corpus object by using tm R package a from messages.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Tm,}
\keyword{corpus}
