% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/is.retweet.r
\name{is.retweet}
\alias{is.retweet}
\title{is.retweet}
\usage{
is.retweet(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Logical value for each message : TRUE/FALSE.
}
\description{
Check if a tweet message have retweet characters
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{retweet}

