% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_channel.r
\name{main_data_channel}
\alias{main_data_channel}
\title{main_data_channel}
\usage{
main_data_channel(x, name, native = F, graph = T, corpus = F, users = T,
  data = T, csv = FALSE, filterdegree = 0, graph_files = FALSE)
}
\arguments{
\item{x}{list  Output obtained by channel_analytic}

\item{name}{character Name of output files.}

\item{native}{logical If channel stream have only native tweets.}

\item{corpus}{logical export csv files.}

\item{users}{logical export users data file.}

\item{csv}{logical Default export is Excel files if csv also desired.}

\item{filterdegree}{numeric Is the degree of filtering.}

\item{graph_files}{logical  Requesting data of centrality measures.}
}
\value{
Data from  \code{channel_analytic} object.
}
\description{
Export result from  \code{channel_analytic} objects.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotation}

