% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_string_vec.r
\name{merge_string_vec}
\alias{merge_string_vec}
\title{merge_string_vec}
\usage{
merge_string_vec(sdata)
}
\arguments{
\item{sdata}{character  char vector}
}
\value{
Return text  merged in one element.
}
\description{
Merge character vec in one string.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{string}
