% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/native_channel.r
\name{native_channel}
\alias{native_channel}
\title{native_channel}
\usage{
native_channel(channel)
}
\arguments{
\item{channel}{data.frame  channel object}
}
\value{
channel_obj data.frame Return a data.frame
}
\description{
extract a channel object only with original tweet
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{channel}

