% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notes_outputs.r
\name{notes_outputs}
\alias{notes_outputs}
\title{notes_outputs}
\usage{
notes_outputs(daily_stat, stat_notes, prefix_file = "notes", na_string = "",
  filecsv = TRUE, html = TRUE, excel = TRUE, daily_unique = FALSE)
}
\arguments{
\item{daily_stat}{data.frame Daily statistics obtained by channel_analytic}

\item{stat_notes}{list Statistical object obtained from stat_notes function.}

\item{prefix_file}{character Prefix for file names outputs}

\item{na_string}{character String value for missing data.}

\item{filecsv}{logical export csv files.}

\item{html}{logical export html sjPlot files.}

\item{excel}{logical export excel files.}

\item{daily_unique}{logical export daily stats of unique.}
}
\value{
A list of daily statistics  data, the unique one and full aggregate stats.
}
\description{
Export result from statistical notes object. Require daily stat object from channel_analytic function.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotation}

