% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/retrieve_annotation_keylist.r
\name{retrieve_annotation_keylist}
\alias{retrieve_annotation_keylist}
\title{retrieve_annotation_keylist}
\usage{
retrieve_annotation_keylist(channel = "allertameteoLIG",
  category = "emergency", target = "hashtag",
  repo = "https://raw.github.com/alfcrisci/twchannels_notations/master/")
}
\arguments{
\item{channel}{character  Name of channel in repository}

\item{category}{character  Annotation category}

\item{target}{character  Target of annotation ( user,hashtag, geonames or other)}

\item{repo}{character  Address of repository where annotation key file are hosted.}
}
\value{
hash_temp  data.frame   A qdap keylist object with key and category
}
\description{
Retrieve a keylist qdap object with  annotations from a github public repository related to a channel.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{match}

