% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweeted_users.r
\name{retweeted_users}
\alias{retweeted_users}
\title{retweeted_users}
\usage{
retweeted_users(x)
}
\arguments{
\item{x}{Character  Messages of channel's tweets}
}
\value{
A vector of retweted users
}
\description{
Extract any user from native retweets from a tweet message
}
\references{
qdapRegex R packages
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{retweet}
\keyword{user,}
