% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_notes.r
\name{stat_notes}
\alias{stat_notes}
\title{stat_notes}
\usage{
stat_notes(x, notes)
}
\arguments{
\item{x}{character  Vector of word to compute statistics of notes}

\item{notes}{qdap hash object  Vector of annotation category}
}
\value{
A list of statistics for each category of annotation present in notes.
}
\description{
Create a list statistical object by using qdap keylist
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotation}

