% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string_occurence.r
\name{string_occurence}
\alias{string_occurence}
\title{string_occurence}
\usage{
string_occurence(x, pattern)
}
\arguments{
\item{x}{character  vector of date correponding to Tweet within unique ID.}

\item{pattern}{Pattern to be search.}
}
\value{
Return a vector with pattern frequency
}
\description{
Extract on a character vector the  frequency of matches relative to textual pattern.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{match}

