% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/string_occurence_vector.r
\name{string_occurence_vector}
\alias{string_occurence_vector}
\title{string_occurence_vector}
\usage{
string_occurence_vector(x, pattern)
}
\arguments{
\item{x}{character  vector.}

\item{pattern}{Pattern to be searched.}
}
\value{
Return a vector with pattern frequency
}
\description{
Extract on a character vector the occurence of textual pattern
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{match}

