% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_at.r
\name{trim_at}
\alias{trim_at}
\title{trim_at}
\usage{
trim_at(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Message of tweet without @ simbol.
}
\description{
Remove at caracter from users mentioned in  tweet message
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{retweet}
