% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_head.r
\name{trim_head}
\alias{trim_head}
\title{trim_head}
\usage{
trim_head(x)
}
\arguments{
\item{x}{Character  Text  of tweet}
}
\value{
Return Text of tweet without head.
}
\description{
Remove head  caracters  before user in  retweeted messages
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{retweet}
