% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim_oddchar.r
\name{trim_oddchar}
\alias{trim_oddchar}
\title{trim_oddchar}
\usage{
trim_oddchar(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Return message of tweet converted in UTF-8.
}
\description{
Remove unicode characters troughout UTF-8 conversion
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{retweet}

