% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_users.r
\name{trim_users}
\alias{trim_users}
\title{trim_users}
\usage{
trim_users(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Return the message of tweet without mentions.
}
\description{
Remove users as mentions from  tweet message
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{retweet}
