% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/channel_analytic.r
\name{channel_analytic}
\alias{channel_analytic}
\title{channel_analytic}
\usage{
channel_analytic(channel_obj, start_date, end_date, Ntop = 11,
  temporal_check = FALSE, Nmin = 25, naming = "DISIT",
  stopword = tm::stopwords("it"))
}
\arguments{
\item{channel_obj}{Data.frame Dataset of tweets}

\item{start_date}{Character Date of   analisys start.}

\item{end_date}{Character Date of  analisys end.}

\item{Ntop}{Integer indicate the maximum number for top statistics}

\item{temporal_check}{Logical indicate if exist time consistency between dates and data}

\item{Nmin}{Integer indicate the minimal data numerosity}

\item{naming}{Character Indicate which naming framework is adopted.}

\item{stopword}{Character stopword set to be use to calculate word frequency matrix. Default italian stopwords of R tm package.}
}
\value{
Return a R list of object useful for channel analisys

channel_stat : channel summaries of following parameters.

*        N_tweets : Number of tweet within unique ID

*        N_retweets  (channel_stat):Number of retweet

*        N_native_tweets (channel_stat):Number of original tweet

*        N_hash  (channel_stat):Number of hashtags detected in channel

*        N_mention (channel_stat): Number of mention detected in channel

*        N_links  (channel_stat):Number of links detected in channel

*        Nuniq_authors (channel_stat):N of unique authors

*        Nuniq_hash (channel_stat): N of unique authors

*        Nuniq_mentions (channel_stat): N of unique authors

*        Nuniq_links (channel_stat):N of unique authors

*        Nfull_retweet (channel_stat):N of comprehensive retweet given by platorm

*        Nfull_retweet_missing (channel_stat):N tweet given without  platorm statistics

*        Nreplies (channel_stat):N of replies

*        Nfavor (channel_stat):N of cumulative favorites given by platorm

table_message :  Frequency data.frame of message

table_hash :  Frequency data.frame of hashtag

table_mentions :  Frequency data.frame of mentions

table_authors :  Frequency data.frame of authors

table_replies :  Frequency data.frame of replies

table_authors_retweeted  :  Frequency data.frame of authors retweeted

table_authors_retweeter :  Frequency data.frame of authors that is a retweeter

rank_authors_retweet :  Frequency data.frame of retweet by using platform data

rank_message_retweet :  Frequency data.frame of message by using platform data

top_message :  TopN messages in channel

top_authors :  TopN authors in channel

top_hash :  TopN hashtag

top_mentions :  TopN user mentioned

top_replies :  TopN user have made replies

top_authors_retweeted :  TopN user that have retweeted in channel

top_authors_retweeter :  TopN user that have made retweet in channel

topfull_authors_retweeted :  TopN author that have retweeted in platform

topfull_message_retweeted :  TopN message that have retweeted in platform

daily_stat :  Daily Temporal data of channel statistic data

authors_date :  DateTime authors activity in channel

links_date :  DateTime authors activity in channel

hash_date :  DateTime hashtag presence in channel

mentions_date :  DateTime mentions presence in channel

unique_message :  Unique message in channel

unique_authors :  Unique authors in channel

unique_hash :  Unique hashtag in channel

unique_mentions :  Unique mentions in channel

unique_authors_retweeted :  Unique retweeted user in channel

unique_authors_retweeter :  Unique retweeter user in channel

uniquefull_authors_retweeted :  Unique retweeted user in platform

uniquefull_message_retweeted :  Unique retweeted message in platform

graph_retweet_df :   Data used for retweet graph

graph_hash_df :  Data used for hashtag graph

graph_mentions_df :  Data for used mention graph

replies_df :  Data for replies

graph_retweet :   Retweet graph object as igraph R object

graph_mentions :   Mention graph object as igraph R object

authors_favorite :   rank of authors favorite

favorite_message_top :   TopN favorite message

channel_corpus :   Tm Corpus of messages without mentions and links and optionally without hashtag

word_freq_matr :   qdap wfm object Word frequency matrix.
}
\description{
Extract many informative stats and object  from a set of tweet messages parsed as channel
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{channel}

