% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_full_notes.r
\name{create_full_notes}
\alias{create_full_notes}
\title{create_full_notes}
\usage{
create_full_notes(obj_notes, label = FALSE)
}
\arguments{
\item{obj_notes}{character Name of channel in repository}

\item{label}{logical  Annotation category}
}
\value{
A qdap keylist object with key and category
}
\description{
Merge keylist  objects with  annotations into a comprehensive one. Need a vector of label note's names.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{match}

