% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_links.r
\name{extract_links}
\alias{extract_links}
\title{extract_links}
\usage{
extract_links(x)
}
\arguments{
\item{x}{Character  Messages of channel's tweets.}
}
\value{
Return the list of mentioned users
}
\description{
Extract url mentioned in a message of tweet.
}
\references{
qdap R packages
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{links}
