% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extract_mentions.r
\name{extract_mentions}
\alias{extract_mentions}
\title{extract_mentions}
\usage{
extract_mentions(x)
}
\arguments{
\item{x}{Character  Messages of channel's tweets.}
}
\value{
Return the list of mentioned users
}
\description{
Extract any mentions, a twitter account mentioned in message, from a tweet.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
qdap R packages
}
\keyword{mentions}

