% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCorpus.r
\name{getCorpus}
\alias{getCorpus}
\title{getCorpus}
\usage{
getCorpus(textVector, hashtag = FALSE)
}
\arguments{
\item{textVector}{Character  Message of tweet}

\item{hashtag}{remove hashtag from corpus}
}
\value{
Return   the tm corpus of message purged by mentions and links.
}
\description{
Creating corpus object by using tm R package a from messages.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Tm,}
\keyword{corpus}

