% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/merge_string_vec.r
\name{merge_string_vec}
\alias{merge_string_vec}
\title{merge_string_vec}
\usage{
merge_string_vec(sdata)
}
\arguments{
\item{sdata}{character  char vector}
}
\value{
Return text  merged in one element.
}
\description{
Merge character vec in one string.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{string}

