% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/retrieve_full_notes.r
\name{retrieve_full_notes}
\alias{retrieve_full_notes}
\title{retrieve_full_notes}
\usage{
retrieve_full_notes(channel, label_notes, category, save = TRUE,
  filerds = "notes.rds")
}
\arguments{
\item{channel}{list  Name of channel in repository}

\item{label_notes}{character  Vector of annotation category}

\item{save}{logical  Save the files in rds format}

\item{filerds}{character  Name of file of rds}

\item{target}{character  Target of annotation ( user, hashtag, geonames or other)}
}
\value{
Return a qdap hash keylist  with key and category
}
\description{
retrieve keylist to perform a build of qdap object from repository.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{annotation}

