% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/trim_hashtags.r
\name{trim_hashtags}
\alias{trim_hashtags}
\title{trim_hashtags}
\usage{
trim_hashtags(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Message of tweets without hashtag
}
\description{
Remove hashtag in a tweet message.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{retweet}

