% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim_urls.r
\name{trim_urls}
\alias{trim_urls}
\title{trim_urls}
\usage{
trim_urls(x)
}
\arguments{
\item{x}{Character  Message of tweet}
}
\value{
Return the message of tweet without links.
}
\description{
Remove links in a tweet message.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{retweet}

